aegis_backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
dnspython
